+++ src/app/layout.tsx
@@ -1,8 +1,6 @@
-'use client';
-
 import type { Metadata } from 'next';
 import React from 'react';
-import { Provider as StoreProvider } from 'react-redux';
+// import { Provider as StoreProvider } from 'react-redux';
 
 import { Footer } from '@/widgets/Footer';
 import { Header } from '@/widgets/Header';
@@ -16,7 +14,7 @@ import { UserProvider } from '@/shared/model/userContext';
 import StyledComponentsRegistry from '@/shared/styles/registry';
 import { GlobalStyles } from '@/shared/styles/reset';
 
-import { store } from './store';
+// import { store } from '@/shared/store/store';
 import { PageContent, StMain } from './styled';
 
 export const metadata: Metadata = {
@@ -30,21 +28,21 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
       <body>
         <StyledComponentsRegistry>
           <GlobalStyles />
-          <StoreProvider store={store}>
-            <PageContent>
-              <UserProvider>
-                <ModalProvider>
-                  <Header />
-                  <StMain>
-                    <Breadcrumbs />
-                    {children}
-                  </StMain>
-                  <Footer />
-                  <Modal />
-                </ModalProvider>
-              </UserProvider>
-            </PageContent>
-          </StoreProvider>
+          {/* <StoreProvider store={store}> */}
+          <PageContent>
+            <UserProvider>
+              <ModalProvider>
+                <Header />
+                <StMain>
+                  <Breadcrumbs />
+                  {children}
+                </StMain>
+                <Footer />
+                <Modal />
+              </ModalProvider>
+            </UserProvider>
+          </PageContent>
+          {/* </StoreProvider> */}
         </StyledComponentsRegistry>
       </body>
     </html>
