--- src/entities/user/model/index.ts
@@ -1,2 +1,0 @@
-export { userSlice, userState, setUser, isLoading, error } from './slice';
-export { fetchGetUser, fetchSignin, fetchSignup } from './sessionThunks';
