+++ src/shared/model/index.ts
@@ -1,1 +1,7 @@
-export { useAppDispatch, useAppSelector } from './hooks';
+import React, { FC } from 'react';
+
+type TLayout = FC<{ children: React.ReactNode }>;
+
+type TDynamicRoute<ParamName extends string> = FC<{ params: { [key in ParamName]: string } }>;
+
+export type { TLayout, TDynamicRoute };
