+++ src/shared/model/typing.ts
