+++ src/shared/store/product/index.ts
