+++ src/shared/store/store.ts
@@ -1,20 +1,16 @@
 import { configureStore, combineReducers } from '@reduxjs/toolkit';
 
-import { userSlice } from '@/entities/user/model';
+import { userSlice } from './user';
 
 const rootReducer = combineReducers({
   user: userSlice.reducer,
 });
 
-export const setupStore = () => {
-  return configureStore({
-    reducer: rootReducer,
-  });
-};
-
-export const store = setupStore();
+export const store = configureStore({
+  reducer: rootReducer,
+});
 
 // Infer the `RootState` and `AppDispatch` types from the store itself
 export type TRootState = ReturnType<typeof rootReducer>;
 // Inferred type: {user: TUserStatus}
-export type TAppDispatch = ReturnType<typeof setupStore>['dispatch'];
+export type TAppDispatch = typeof store.dispatch;
