+++ src/shared/store/user/index.ts
@@ -1,0 +1,2 @@
+export { userSlice, getUser, setUser, isLoading, error } from './slice';
+export { fetchGetUser, fetchSignin, fetchSignup } from './thunks';
